Read data from the board in a separate thread
With GGB6, reading data on-demand can result in stale data being
returned. This is probably due to the way GGB6 continuously sends data
multiple times a second. If that data isn't read quickly enough, it
could start filling up the buffer. Subsequent reads would then pull the
data from the buffer, which is likely be outdated. Fix this by creating
a thread that continuously reads data from the board. The read() method
is also adjusted to simply returns data from the latest read instead of
reading from the board directly.