Add primitive "init"
The primitive can be used to manually load the daemon. It also accepts a
parameter that allows the user to choose whether or not to use the gogo6
version of the daemon. Users are not required to call this primitive
before using other primitives; the daemon will still be loaded
automatically if this primitive isn't called. The old daemon will be
used in that case.